requirements.txt
# Flow Farm - 手机流量农场自动化系统依赖

# ============ 核心依赖 ============
# GUI框架 - tkinter (Python内置，无需安装)
Pillow>=10.1.0                   # 图像处理

# 设备控制与自动化
uiautomator2>=2.16.23            # Android自动化
pure-python-adb>=0.3.0.dev0     # ADB接口

# 网络请求
requests>=2.31.0                 # HTTP客户端
urllib3>=2.0.7                   # URL处理

# 加密安全
cryptography>=41.0.7             # 加密库
bcrypt>=4.1.2                    # 密码哈希

# 配置管理
python-dotenv>=1.0.0             # 环境变量管理（可选）

# 系统工具
psutil>=5.9.6                    # 系统监控
colorama>=0.4.6                  # 颜色输出

# ============ 开发工具 ============
# 代码质量
black>=23.11.0                   # 代码格式化
pylint>=3.0.3                    # 代码检查
mypy>=1.7.1                      # 类型检查

# 测试框架
pytest>=7.4.3                    # 单元测试
pytest-cov>=4.1.0                # 代码覆盖率

# ============ 构建打包 ============
PyInstaller>=6.2.0               # 程序打包

# ============ 平台特定 ============
pywin32>=306; sys_platform == "win32"                    # Windows特定
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"  # macOS特定

# ============ 可选增强 ============
# 以下为可选依赖，根据需要安装
# opencv-python>=4.8.1.78        # 图像识别（高级功能）
# numpy>=1.24.4                   # 数值计算（图像处理需要）
# pandas>=2.1.4                   # 数据分析（统计功能）
# tqdm>=4.66.1                    # 进度条显示
# schedule>=1.2.0                 # 任务调度（高级任务管理）
threading                        # 多线程（内置）
asyncio                          # 异步IO（内置）
concurrent.futures               # 并发执行（内置）

# 打包构建
PyInstaller>=6.2.0               # Python打包工具
auto-py-to-exe>=2.42.0          # GUI打包工具

# XML处理
lxml>=4.9.3                      # XML解析器
beautifulsoup4>=4.12.2          # HTML/XML解析

# 时间处理
arrow>=1.3.0                     # 时间处理库
python-dateutil>=2.8.2          # 日期工具

# 文件操作
pathlib                          # 路径处理（内置）
shutil                           # 文件操作（内置）
zipfile                          # 压缩文件（内置）

# 正则表达式和字符串
re                               # 正则表达式（内置）
string                           # 字符串工具（内置）

# JSON处理
json                             # JSON处理（内置）
ujson>=5.8.0                    # 快速JSON库

# 进程管理
subprocess                       # 子进程（内置）

# 平台相关
platform                        # 平台检测（内置）
os                              # 操作系统接口（内置）
sys                             # 系统参数（内置）

# 可选依赖（根据需要安装）
# PyQt5>=5.15.10                # PyQt5 GUI框架（备选）
# PyQt6>=6.6.0                  # PyQt6 GUI框架（备选）
# kivy>=2.2.2                   # 移动应用框架（备选）
